+++ app/src/main/java/org/privacymatters/safespace/cameraUtils/CameraSwitch.kt
@@ -1,0 +1,7 @@
+package org.privacymatters.safespace.cameraUtils
+
+interface CameraSwitch {
+
+    fun switchMode(mode: String)
+    fun switchCamera(mode: String)
+}
