+++ app/src/main/res/values/strings.xml
@@ -12,11 +12,14 @@
     <string name="q720p">720p</string>
     <string name="fps30">30 FPS</string>
     <string name="fps60">60 FPS</string>
+    <string name="video_quality">Video Quality</string>
+    <string name="video_fps">Video FPS</string>
     <string name="take_photo">Take Photo</string>
     <string name="flash_toggle">Flash</string>
     <string name="start_capture">Start Capture</string>
     <string name="stop_capture">Stop Capture</string>
     <string name="video_error">Internal Error, could not save video</string>
+    <string name="video_timer">00:00</string>
     <string name="permission_denied">Permission request denied</string>
     <string name="import_files_error">Some files were not imported, please try again</string>
     <string name="create_folder_invalid_error">Only letters and numbers allowed </string>
@@ -85,4 +88,6 @@
     <!-- TODO: Remove or change this placeholder text -->
     <string name="image_capture_fragment">Image Capture</string>
     <string name="video_capture_fragment">Video Capture</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
